requirements/base.txt
Django==5.0.1
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-health-check==3.18.1
drf-nested-routers==0.93.5
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
six==1.16.0
django-filter==23.5
drf-yasg==1.21.7
django-countries==7.5.1
django-localflavor==4.0
python-dateutil==2.9.0.post0
